Not working right now

+++ Tidying2/state_michael2/schedule.h
@@ -16,6 +16,6 @@ class schedule{
 };
 
 
-bool wait (int instance, int delayTime);
+bool wait (int instance, unsigned long delayTime);
 
 #endif
